+++ src/main.js
@@ -13,7 +13,7 @@ const createCharactersHtml = (characters) =>{
   })
   return characterName
 }
-const characterNotFound = (characters) => {
+const characterNotFound = () => {
   let characterDataNotFound =
     `<div class="characterNotFound">
       <h3 class="notFound">No results for this search</h3>
@@ -52,29 +52,35 @@ inputs.forEach(input =>{
 // esconder los ... y mostrar characters
 const dotButton = document.getElementById("dotsButton");
 
-dotButton.addEventListener("click", () => {
-  document.getElementById("dotsButton").hidden = true;
-  document.getElementById("charactersButton").classList.remove("inactive");
-})
+if (dotButton){
+  dotButton.addEventListener("click", () => {
+    document.getElementById("dotsButton").hidden = true;
+    document.getElementById("charactersButton").classList.remove("inactive");
+  })
+}
 
 // al hacer click en characters, se esconde home y se muestra el nav con los personajes (nombre e imagen)
 const characterButton = document.getElementById("charactersButton");
 
-characterButton.addEventListener("click", async() => {
-  document.getElementById("home").hidden = true;
-  document.getElementById("navSearcher").classList.remove("inactive");
-  let data = await getAllCharacters();
-  window.localStorage.setItem('Character', JSON.stringify(data));
-  const charactersHTML = createCharactersHtml(data);
-  document.getElementById("results").innerHTML = charactersHTML
-})
+if (characterButton){
+  characterButton.addEventListener("click", async() => {
+    document.getElementById("home").hidden = true;
+    document.getElementById("navSearcher").classList.remove("inactive");
+    let data = await getAllCharacters();
+    window.localStorage.setItem('Character', JSON.stringify(data));
+    const charactersHTML = createCharactersHtml(data);
+    document.getElementById("results").innerHTML = charactersHTML
+  })
+}
 
 // recargar página una vez que se presiona el logo y va a home
 const logoButtonNav = document.getElementById("logoButtonNav");
 
-logoButtonNav.addEventListener("click", () => {
-  window.location.reload(logoButtonNav)
-})
+if (logoButtonNav){
+  logoButtonNav.addEventListener("click", () => {
+    window.location.reload(logoButtonNav)
+  })
+}
 
 /* Sort function */
 const sortABtn = document.querySelector('#a-z');
@@ -89,15 +95,18 @@ export function sortingCharacters (charactersList){
   return newCharactersSorted
 }
 
-sortABtn.onclick = () => {
-  const newCharacters = JSON.parse(window.localStorage.getItem('Character'));
-  const charactersHTML = createCharactersHtml(sortingCharacters(newCharacters));
-  document.getElementById("results").innerHTML = charactersHTML
+if (sortABtn){
+  sortABtn.onclick = () => {
+    const newCharacters = JSON.parse(window.localStorage.getItem('Character'));
+    const charactersHTML = createCharactersHtml(sortingCharacters(newCharacters));
+    document.getElementById("results").innerHTML = charactersHTML
+  }
 }
 
-sortZBtn.onclick = () => {
-  console.log('Ordename z - a')
-  const newCharacters = JSON.parse(window.localStorage.getItem('Character'));
-  const charactersHTML = createCharactersHtml(sortingCharacters(newCharacters).reverse());
-  document.getElementById("results").innerHTML = charactersHTML
+if (sortZBtn){
+  sortZBtn.onclick = () => {
+    const newCharacters = JSON.parse(window.localStorage.getItem('Character'));
+    const charactersHTML = createCharactersHtml(sortingCharacters(newCharacters).reverse());
+    document.getElementById("results").innerHTML = charactersHTML
+  }
 }
